# ARROW-12190: [Rust][DataFusion] Implement parallel / partitioned hash join