--- a/rust/datafusion/src/physical_plan/planner.rs
+++ b/rust/datafusion/src/physical_plan/planner.rs
@@ -20,7 +20,8 @@
 use std::sync::Arc;
 
 use super::{
-    aggregates, empty::EmptyExec, expressions::binary, functions, udaf, union::UnionExec,
+    aggregates, empty::EmptyExec, expressions::binary, functions,
+    hash_join::PartitionMode, udaf, union::UnionExec,
 };
 use crate::error::{DataFusionError, Result};
 use crate::execution::context::ExecutionContextState;
@@ -344,12 +345,22 @@ impl DefaultPhysicalPlanner {
                     JoinType::Left => hash_utils::JoinType::Left,
                     JoinType::Right => hash_utils::JoinType::Right,
                 };
+                let left_expr = keys.iter().map(|x| col(&x.0)).collect();
+                let right_expr = keys.iter().map(|x| col(&x.1)).collect();
 
+                // Use hash partition by defualt to parallelize hash joins
                 Ok(Arc::new(HashJoinExec::try_new(
-                    left,
-                    right,
+                    Arc::new(RepartitionExec::try_new(
+                        left,
+                        Partitioning::Hash(left_expr, ctx_state.config.concurrency),
+                    )?),
+                    Arc::new(RepartitionExec::try_new(
+                        right,
+                        Partitioning::Hash(right_expr, ctx_state.config.concurrency),
+                    )?),
                     &keys,
                     &physical_join_type,
+                    PartitionMode::Partitioned,
                 )?))
             }
             LogicalPlan::EmptyRelation {