--- a/rust/datafusion/src/physical_plan/planner.rs
+++ b/rust/datafusion/src/physical_plan/planner.rs
@@ -345,23 +345,33 @@ impl DefaultPhysicalPlanner {
                     JoinType::Left => hash_utils::JoinType::Left,
                     JoinType::Right => hash_utils::JoinType::Right,
                 };
-                let left_expr = keys.iter().map(|x| col(&x.0)).collect();
-                let right_expr = keys.iter().map(|x| col(&x.1)).collect();
+                if ctx_state.config.concurrency > 1 {
+                    let left_expr = keys.iter().map(|x| col(&x.0)).collect();
+                    let right_expr = keys.iter().map(|x| col(&x.1)).collect();
 
-                // Use hash partition by defualt to parallelize hash joins
-                Ok(Arc::new(HashJoinExec::try_new(
-                    Arc::new(RepartitionExec::try_new(
+                    // Use hash partition by defualt to parallelize hash joins
+                    Ok(Arc::new(HashJoinExec::try_new(
+                        Arc::new(RepartitionExec::try_new(
+                            left,
+                            Partitioning::Hash(left_expr, ctx_state.config.concurrency),
+                        )?),
+                        Arc::new(RepartitionExec::try_new(
+                            right,
+                            Partitioning::Hash(right_expr, ctx_state.config.concurrency),
+                        )?),
+                        &keys,
+                        &physical_join_type,
+                        PartitionMode::Partitioned,
+                    )?))
+                } else {
+                    Ok(Arc::new(HashJoinExec::try_new(
                         left,
-                        Partitioning::Hash(left_expr, ctx_state.config.concurrency),
-                    )?),
-                    Arc::new(RepartitionExec::try_new(
                         right,
-                        Partitioning::Hash(right_expr, ctx_state.config.concurrency),
-                    )?),
-                    &keys,
-                    &physical_join_type,
-                    PartitionMode::Partitioned,
-                )?))
+                        &keys,
+                        &physical_join_type,
+                        PartitionMode::MergeLeft,
+                    )?))
+                }
             }
             LogicalPlan::EmptyRelation {
                 produce_one_row,