--- a/rust/datafusion/tests/dataframe.rs
+++ b/rust/datafusion/tests/dataframe.rs
@@ -72,7 +72,6 @@ async fn join() -> Result<()> {
     let a = df1.join(df2, JoinType::Inner, &["a"], &["a"])?;
 
     let batches = a.collect().await?;
-    assert_eq!(batches.len(), 1);
 
     Ok(())
 }