--- a/rust/datafusion/tests/dataframe.rs
+++ b/rust/datafusion/tests/dataframe.rs
@@ -71,7 +71,7 @@ async fn join() -> Result<()> {
 
     let a = df1.join(df2, JoinType::Inner, &["a"], &["a"])?;
 
-    let batches = a.collect().await?;
+    let _ = a.collect().await?;
 
     Ok(())
 }