@Dandandan 2021-04-04
@jorgecarleitao I am wondering if you have found a more performant implementation or ideas for the `concat` kernel in your `arrow2` branch?
In this case (even though the extra parallelism has better performance and better scalability), through CoalesceBatches it is adding significant overhead, but looking at the implementation think there are probably quite some opportunities to improve performance there - especially for the concat case there seems too much work currently done in the `MutableArrayData` and creating intermediate `Vec`s. etc.