@andygrove 2021-04-04
This is looking really good. I agree with your suggestion about buffering data and only emitting batches when they reach a target size.