@Dandandan 2021-04-06
@alamb added one test with multiple partitions - maybe later I can add some more tests. I also tried running the TPC-H benchmarks, which return the same results / failures as on master.